started code cleanup, db schema remodeling
