+++ src/app/routers/login.py
@@ -3,7 +3,8 @@ from pycountry import countries
 from pydantic import BaseModel
 
 from ..database.authentication import Authentication
-from .authentication import User, authenticate
+from .authentication import authenticate
+from ..database.models.models import User
 
 
 
@@ -33,16 +34,8 @@ async def is_username(username: str, user: User = Depends(authenticate)):
 
     return Authentication.is_username(username)
 
-class New_User(BaseModel):
-    username: str
-    password: str
-    company: str
-    country: str
-    address: str
-    access_lvl: int
-
 @router.post("/signup")
-async def signup(new_user: New_User, user: User = Depends(authenticate)):
+async def signup(new_user: User, user: User = Depends(authenticate)):
     if user["access_lvl"] != 3 and user["access_lvl"] != 4:
         raise HTTPException(status_code=403, detail="Insufficient authorization level!")
