+++ src/dockerfile
@@ -4,4 +4,6 @@ COPY . .
 
 RUN pip install -r requirements.txt
 
+ENV mode=DOCKER
+
 CMD [ "uvicorn", "--host", "0.0.0.0", "app.main:app" ]
